data/requirements.txt
absl-py==0.10.0
alabaster==0.7.12
ampligraph==1.3.2
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.1.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
beautifultable==1.0.0
black==20.8b1
bleach==3.2.1
blis==0.4.1
bs4==0.0.1
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
commonmark==0.9.1
constantly==15.1.0
contextvars==2.4
cryptography==3.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dask==2.26.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
distributed==2.26.0
docutils==0.16
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
flake8==3.8.4
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==3.1.0
HeapDict==1.0.1
html5lib==1.1
hyperlink==20.0.1
idna==2.10
imagesize==1.2.0
immutables==0.14
importlib-metadata==1.7.0
incremental==17.5.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itemadapter==0.1.0
itemloaders==1.0.2
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
latexcodec==2.0.1
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
money-parser==0.0.1
msgpack==1.0.0
multiprocess==0.70.10
munkres==1.1.4
murmurhash==1.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.22
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
neo4j==4.1.1
nest-asyncio==1.4.1
networkx==2.3
nltk==3.5
notebook==6.1.4
numba==0.51.2
numbskull==0.1.1
numpy==1.18.5
opt-einsum==3.3.0
oset==0.1.3
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
Protego==0.1.16
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.23.0
pybtex-docutils==0.2.2
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyflakes==2.2.0
Pygments==2.7.1
PyHamcrest==2.0.2
pyjarowinkler==1.8
pynvim==0.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrallel.lib==0.0.8
pyrsistent==0.17.3
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
queuelib==1.5.0
rdflib==5.0.0
recommonmark==0.6.0
regex==2020.9.27
requests==2.24.0
rltk==2.0.0a15
scikit-learn==0.21.3
scipy==1.5.4
Scrapy==2.3.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
snorkel @ file:///Users/rogerlin/PycharmProjects/USCKnowledgeGraph/Snorkel/snorkel-0.7.0-beta
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
SPARQLWrapper==1.8.5
Sphinx==2.4.4
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==1.0.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.19
srsly==1.0.2
tblib==1.7.0
tensorboard==1.15.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==7.4.1
toml==0.10.1
toolz==0.10.0
torch==1.6.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
treedlib==0.1.3
Twisted==20.3.0
typed-ast==1.4.1
typing==3.7.4.3
typing-extensions==3.7.4.3
ujson==3.1.0
urllib3==1.25.10
w3lib==1.22.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zict==2.0.0
zipp==3.1.0
zope.interface==5.1.0
